+++ com.onesignal.unity.core/Editor/Platform/OneSignalNative.cs
@@ -41,13 +41,18 @@ namespace OneSignalSDK {
         public override event InAppMessageLifecycleDelegate InAppMessageWillDismiss;
         public override event InAppMessageLifecycleDelegate InAppMessageDidDismiss;
         public override event InAppMessageActionDelegate InAppMessageTriggeredAction;
-        public override event StateChangeDelegate<PermissionState> PermissionStateChanged;
+        public override event StateChangeDelegate<NotificationPermission> NotificationPermissionChanged;
         public override event StateChangeDelegate<PushSubscriptionState> PushSubscriptionStateChanged;
         public override event StateChangeDelegate<EmailSubscriptionState> EmailSubscriptionStateChanged;
         public override event StateChangeDelegate<SMSSubscriptionState> SMSSubscriptionStateChanged;
 
         public override bool PrivacyConsent { get; set; }
         public override bool RequiresPrivacyConsent { get; set; }
+        
+        public override NotificationPermission NotificationPermission { get; }
+        public override PushSubscriptionState PushSubscriptionState { get; }
+        public override EmailSubscriptionState EmailSubscriptionState { get; }
+        public override SMSSubscriptionState SMSSubscriptionState { get; }
 
         public override void Initialize(string appId) {
             if (string.IsNullOrEmpty(appId)) {