+++ com.onesignal.unity.core/Runtime/Models/NotificationPermission.cs
@@ -25,32 +25,24 @@
  * THE SOFTWARE.
  */
 
-using System;
-
 namespace OneSignalSDK {
     /// <summary>
-    /// Status of ability to send push notification as determined by the current user
+    /// Status of ability to send push notifications to the current device as determined by its user
     /// </summary>
     public enum NotificationPermission {
         /// <summary>The user has not yet made a choice regarding whether your app can show notifications.</summary>
-        NotDetermined,
+        NotDetermined = 0,
 
         /// <summary>The application is not authorized to post user notifications.</summary>
         Denied,
 
         /// <summary>The application is authorized to post user notifications.</summary>
         Authorized,
-        
+
         /// <summary>The application is only authorized to post Provisional notifications (direct to history)</summary>
         Provisional,
-        
+
         /// <summary>The application is authorized to send notifications for 8 hours. Only used by App Clips.</summary>
         Ephemeral
-    }
-
-    [Serializable] public sealed class PermissionState {
-        public bool hasPrompted;
-        public NotificationPermission status;
-        public bool provisional;
     }
 }