+++ OneSignalExample/Assets/OneSignal/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Add missing getters for permission and subscription states
+### Changed
+- Removed `PermissionState` in favor of `NotificationPermission` enum
+  - Renamed `PermissionStateChanged` event to `NotificationPermissionChanged`
 
 ## [3.0.0-beta.2]
 ### Fixed